+++ src/main/java/dev/lambdaurora/auroraslanterns/block/AmethystLanternBlock.java
@@ -9,31 +9,35 @@
 
 package dev.lambdaurora.auroraslanterns.block;
 
-import dev.lambdaurora.auroraslanterns.AurorasLanterns;
+import com.mojang.serialization.MapCodec;
 import dev.lambdaurora.auroraslanterns.AurorasLanternsRegistry;
-import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.minecraft.core.BlockPos;
-import net.minecraft.resources.Identifier;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.LanternBlock;
 import net.minecraft.world.level.block.state.BlockState;
+import org.jetbrains.annotations.NotNull;
 
 /**
  * Represents an amethyst lantern block.
  *
  * @author LambdAurora
- * @version 1.0.0
+ * @version 1.1.0
  * @since 1.0.0
  */
 public class AmethystLanternBlock extends LanternBlock {
-	public static final Identifier BLOCK_TEXTURE = AurorasLanterns.id("block/amethyst_lantern");
-	public static final Identifier HANGING_MODEL = AurorasLanterns.id("block/hanging_amethyst_lantern");
+	public static final MapCodec<AmethystLanternBlock> CODEC = simpleCodec(AmethystLanternBlock::new);
+
 	public static final int EFFECT_RADIUS = 32;
 
-	public AmethystLanternBlock() {
-		super(FabricBlockSettings.copyOf(Blocks.LANTERN).luminance(14));
+	public AmethystLanternBlock(Properties properties) {
+		super(properties);
+	}
+
+	@SuppressWarnings({"rawtypes", "unchecked"})
+	@Override
+	public @NotNull MapCodec<LanternBlock> codec() {
+		return (MapCodec) CODEC;
 	}
 
 	/* Visual */