+++ src/main/java/dev/lambdaurora/auroraslanterns/block/RedstoneWallLanternBlock.java
@@ -9,6 +9,7 @@
 
 package dev.lambdaurora.auroraslanterns.block;
 
+import com.mojang.serialization.MapCodec;
 import dev.lambdaurora.auroraslanterns.block.behavior.RedstoneLanternBehavior;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -18,19 +19,27 @@ import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
+import org.jetbrains.annotations.NotNull;
 
-@SuppressWarnings("deprecation")
 public class RedstoneWallLanternBlock extends WallLanternBlock<RedstoneLanternBlock> {
+	public static final MapCodec<? extends RedstoneWallLanternBlock> CODEC
+			= makeCodec(RedstoneLanternBlock.class, RedstoneWallLanternBlock::new);
+
 	private final RedstoneLanternBehavior behavior = new RedstoneLanternBehavior(state -> state.get(FACING));
 
 	public RedstoneWallLanternBlock(RedstoneLanternBlock lantern) {
 		super(lantern);
 	}
 
+	@Override
+	protected @NotNull MapCodec<? extends RedstoneWallLanternBlock> codec() {
+		return CODEC;
+	}
+
 	/* Updates */
 
 	@Override
-	public void neighborChanged(BlockState state, Level world, BlockPos pos, Block block, BlockPos fromPos, boolean notify) {
+	protected void neighborChanged(BlockState state, Level world, BlockPos pos, Block block, BlockPos fromPos, boolean notify) {
 		super.neighborChanged(state, world, pos, block, fromPos, notify);
 		this.behavior.neighborChanged(state, world, pos);
 	}
@@ -38,7 +47,7 @@ public class RedstoneWallLanternBlock extends WallLanternBlock<RedstoneLanternBl
 	/* Ticking */
 
 	@Override
-	public void tick(BlockState state, ServerLevel world, BlockPos pos, RandomSource random) {
+	protected void tick(BlockState state, ServerLevel world, BlockPos pos, RandomSource random) {
 		super.tick(state, world, pos, random);
 		this.behavior.scheduledTick(state, world, pos);
 	}
@@ -46,12 +55,12 @@ public class RedstoneWallLanternBlock extends WallLanternBlock<RedstoneLanternBl
 	/* Redstone */
 
 	@Override
-	public int getSignal(BlockState state, BlockGetter world, BlockPos pos, Direction direction) {
+	protected int getSignal(BlockState state, BlockGetter world, BlockPos pos, Direction direction) {
 		return this.behavior.getWeakRedstonePower(state, world, pos, direction);
 	}
 
 	@Override
-	public int getDirectSignal(BlockState state, BlockGetter world, BlockPos pos, Direction direction) {
+	protected int getDirectSignal(BlockState state, BlockGetter world, BlockPos pos, Direction direction) {
 		return this.behavior.getStrongRedstonePower(state, world, pos, direction);
 	}
 }