+++ src/main/java/dev/lambdaurora/auroraslanterns/resource/AurorasLanternsStaticDatagen.java
@@ -18,14 +18,14 @@ import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricAdvancementProvider;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricBlockLootTableProvider;
 import net.minecraft.advancements.Advancement;
-import net.minecraft.advancements.FrameType;
-import net.minecraft.advancements.RequirementsStrategy;
-import net.minecraft.advancements.critereon.KilledTrigger;
+import net.minecraft.advancements.AdvancementHolder;
+import net.minecraft.advancements.AdvancementType;
+import net.minecraft.core.HolderLookup;
 import net.minecraft.network.chat.Text;
 import net.minecraft.resources.Identifier;
-import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.Blocks;
 
+import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 
 public final class AurorasLanternsStaticDatagen implements DataGeneratorEntrypoint {
@@ -37,8 +37,8 @@ public final class AurorasLanternsStaticDatagen implements DataGeneratorEntrypoi
 	}
 
 	private static class LootDataProvider extends FabricBlockLootTableProvider {
-		public LootDataProvider(FabricDataOutput output) {
-			super(output);
+		public LootDataProvider(FabricDataOutput output, CompletableFuture<HolderLookup.Provider> registryLookup) {
+			super(output, registryLookup);
 		}
 
 		@Override
@@ -49,30 +49,14 @@ public final class AurorasLanternsStaticDatagen implements DataGeneratorEntrypoi
 	}
 
 	private static class AdvancementProvider extends FabricAdvancementProvider {
-		public AdvancementProvider(FabricDataOutput output) {
-			super(output);
+		public AdvancementProvider(FabricDataOutput output, CompletableFuture<HolderLookup.Provider> registryLookup) {
+			super(output, registryLookup);
 		}
 
 		@Override
-		public void generateAdvancement(Consumer<Advancement> consumer) {
-			Advancement root = Advancement.Builder.advancement()
-					.display(
-							Items.MAP,
-							Text.translatable("advancements.adventure.root.title"),
-							Text.translatable("advancements.adventure.root.description"),
-							new Identifier("textures/gui/advancements/backgrounds/adventure.png"),
-							FrameType.TASK,
-							false,
-							false,
-							false
-					)
-					.requirements(RequirementsStrategy.OR)
-					.addCriterion("killed_something", KilledTrigger.TriggerInstance.playerKilledEntity())
-					.addCriterion("killed_by_something", KilledTrigger.TriggerInstance.entityKilledPlayer())
-					.build(new Identifier(Identifier.DEFAULT_NAMESPACE, "adventure/root"));
-
+		public void generateAdvancement(HolderLookup.Provider registryLookup, Consumer<AdvancementHolder> consumer) {
 			consumer.accept(Advancement.Builder.advancement()
-					.parent(root)
+					.parent(new AdvancementHolder(Identifier.ofDefault("adventure/root"), null))
 					.display(
 							Blocks.LANTERN,
 							Text.translatable(
@@ -82,12 +66,12 @@ public final class AurorasLanternsStaticDatagen implements DataGeneratorEntrypoi
 									"advancements.%s.adventure.wall_lantern_bonk.description".formatted(AurorasLanterns.NAMESPACE)
 							),
 							null,
-							FrameType.TASK,
+							AdvancementType.TASK,
 							true,
 							true,
 							false
 					)
-					.addCriterion("bonk", WallLanternBonkTrigger.TriggerInstance.bonk(null))
+					.addCriterion("bonk", WallLanternBonkTrigger.TriggerInstance.bonk())
 					.build(AurorasLanterns.id("adventure/wall_lantern_bonk"))
 			);
 		}